backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
alembic==1.15.2
pydantic>=2.12.2
pydantic-settings==2.8.1
python-dotenv==1.1.0
openai-agents==0.13.3
openai>=1.76.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
authlib==1.3.2
psycopg2-binary==2.9.10
